Update addrecordview without converter and update sheet in home view

+++ GlucoseTracker/GlucoseTracker/Model/GlucoseData.swift
@@ -11,11 +11,11 @@ import SwiftData
 @Model
 class GlucoseData {
     var date: Date
-    var amount: Double
+    var amount: Int
     var type: String
     var notes: String
     
-    init(date: Date, amount: Double, type: String, notes: String) {
+    init(date: Date, amount: Int, type: String, notes: String) {
         self.date = date
         self.amount = amount
         self.type = type
